remove show-tree, update tree-status
