+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
   "tests/acc_forwarder",
   "tests/bgtask_creator",
   "tests/load_generation",
-  "tests/show-tree",
   "tests/tree-status",
   "tests/txn_forwarder",
 ]
