--- tests/show-tree/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "show-tree"
-version = "0.6.9"
-edition = "2021"
-publish = false
-
-[dependencies]
-anchor-client = "0.26.0"
-anyhow = "1.0.71"
-bs58 = "0.4.0"
-clap = { version = "4.1.4", features = ["derive"] }
-futures = "0.3.28"
-solana-client = "=1.14.15"
-solana-sdk = "=1.14.15"
-solana-transaction-status = "=1.14.15"
-spl-account-compression = { version = "0.1.10", features = ["no-entrypoint"] }
-spl-noop = { version ="0.1.3", features = ["no-entrypoint"] }
-thiserror = "1.0.31"
-tokio = { version = "1.23.0", features = ["macros", "rt-multi-thread", "sync", "time"] }
-txn_forwarder = { path = "../txn_forwarder" }
