+++ tests/tree-status/Cargo.toml
@@ -9,11 +9,11 @@ publish = false
 [dependencies]
 anchor-client = "0.26.0"
 anyhow = "1.0.70"
-async-recursion = "1.0.4"
 bs58 = "0.4.0"
 clap = { version = "4.1.4", features = ["derive"] }
 digital_asset_types = { path = "../../digital_asset_types", features = ["json_types", "sql_types"] }
 flatbuffers = "23.1.21"
+futures = "0.3.28"
 hex = "0.4.3"
 sea-orm = { version = "0.10.6", features = ["macros", "runtime-tokio-rustls", "sqlx-postgres", "with-chrono", "mock"] }
 solana-client = "=1.14.15"
@@ -23,5 +23,5 @@ spl-account-compression = { version = "0.1.10", features = ["no-entrypoint"] }
 spl-noop = { version ="0.1.3", features = ["no-entrypoint"] }
 sqlx = { version = "0.6.2", features = ["macros", "runtime-tokio-rustls", "postgres", "uuid", "offline", "json"] }
 thiserror = "1.0.31"
-tokio = { version = "1.23.0", features = ["fs", "macros", "rt-multi-thread"] }
+tokio = { version = "1.23.0", features = ["fs", "macros", "rt-multi-thread", "sync", "time"] }
 txn_forwarder = { path = "../txn_forwarder" }
